Copying student submissions to grading directory
and checking for submitted files. If submitted file are not found, check
whether the student has submitted a file with an alias name. If a file
with alias is found it is renamed in to the file the student supposed to
submit.

Renaming might be needed to be restricted to the situation where we are
providing a Makefile.