Creating grading log files
Started creating grading log files. Discovered a bug. I clone student
repos into grading directory and re-name files that match file aliases
to the names what they are supposed to be. When I pull into grading
directory for the second time, since an old file with the correct name
now exists in the student repo in grading directory, the updated file
goes undetected and grading is performed on the old file. Need to find a
way to remove changed files before pulling again or something...