Entering grades to gradebook.
Working on recording marks in the gradebook. There is a problem in the
grading work floor. I check for existence of required file for the whole
assignment and if some files are missing I abort grading for that
student. Instead, I need to check required files for each problem of an
assignment and abort only the problems that do not have all the required
files, but continue grading on the rest of the problems.