Moved gradebook headers to AgGlobals, Implemented checking dependencies.
Implemented checking for success of dependent problems before evaluating
a problem. This part has not been tested at all.

When implementing this part, enforced the restriction of problem numbers
having to be integers. Also, a problem with numbe n can only be
dependent on problems p such that p < n.