In Input.py handled a error case when loading an input config.
Started updating Problem.py to grade output produced in files. Added a
method to Input.py to support this enhancement.

To be done: Consolidate the matching ratios of standard output and files
output for a single set of inputs to get a ratio for that input. Then
consolidate all the mathing ratios of all the inputs to decide the final
grade for the problem.