Did some modifications:
Updated problem configuration to have a ignore_spaces flag to be used in
grading. If the flag is set to 'yes' then when comparing student output
to reference output, mismatches due to missing of additional spaces in
student output are ignored.

Add a weight attribute to input. Using this different weights can be
assigned to different inputs. These weights are used in grading. A
problem having a poor match for the output produced by input with a high
weight looses more marks.

Removed most of the debugging output to the screen.

I think now we are ready for a larger test run.