Completed adding a fixed seed to some designated files.
Added "_17_random" field to the Problem class, which adds random = field
to the problem configuration file.

Updated the problem class to pass the new problem configuratin file.

Now when compiling code, it searches for srand() in the designated
random files and changes the seed to the designated seed before
compiling. After compiling it changes the files back to as it is.